Version 1.0.0 Build 1.0.0 - Initial Upload
Populagram is an application that was designed to demonstrate my
knowledge of CoreData, GCD, NSOperationQueues, UIKit, MapKit and
CoreLocation by using Instagram's Popular API.

I also used this application to test the power of CoreGraphics in a
UITableView. I wanted to see if it was actually faster to draw the
cell's content over using multiple subview. While the results were
pretty much similar on newer devices (i.e. iPhone 5) I found that this
method was actually faster for older devices such as an iPod Touch 4.

* You need to provide your own Instagram client id in order to use the
app *